[chore] Flytte og døpe om filer
