--- src/digisos/informasjon/personopplysninger.less
@@ -1,25 +1,0 @@
-.personopplysning_info {
-    padding: 3rem 2rem 0 2rem;
-    max-width: 600px;
-
-    @media @mobile {
-        padding-top: 3rem;
-        padding-left: 1rem;
-        padding-right: 1rem;
-        padding-bottom: 1rem;
-    }
-
-    h3 {
-        &:extend(.typo-systemtittel);
-    }
-
-    h4 {
-        &:extend(.typo-undertittel);
-        margin-top: 2rem;
-    }
-
-    h5 {
-        &:extend(.typo-element);
-        margin-top: 1rem;
-    }
-}
