+++ src/digisos/informasjon/personopplysninger/PersonopplysningerKortfattet.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import {FormattedMessage} from "react-intl";
 import {useDispatch} from "react-redux";
-import BehandlingAvPersonopplysningerModal from "./BehandlingAvPersonopplysningerModal";
-import {getContextPathForStaticContent} from "../../configuration";
-import {visSamtykkeInfo} from "../redux/soknad/soknadActions";
-import {LinkButton} from "../../nav-soknad/components/linkButton/LinkButton";
+import PersonopplysningerUtfyllendeModal from "./PersonopplysningerUtfyllendeModal";
+import {getContextPathForStaticContent} from "../../../configuration";
+import {visSamtykkeInfo} from "../../redux/soknad/soknadActions";
+import {LinkButton} from "../../../nav-soknad/components/linkButton/LinkButton";
 import {BodyLong, Heading} from "@navikt/ds-react";
 import styled from "styled-components";
 
@@ -57,7 +57,7 @@ const AndreSporsmalTekst = styled.div`
     margin-top: 1rem;
 `;
 
-const Personopplysninger = () => {
+const PersonopplysningerKortfattet = () => {
     const dispatch = useDispatch();
 
     return (
@@ -133,9 +133,9 @@ const Personopplysninger = () => {
                     <FormattedMessage id="informasjon.tekster.personopplysninger.sporsmal" />
                 </BodyLong>
             </AndreSporsmalTekst>
-            <BehandlingAvPersonopplysningerModal />
+            <PersonopplysningerUtfyllendeModal />
         </PersonopplysningPanel>
     );
 };
 
-export default Personopplysninger;
+export default PersonopplysningerKortfattet;
