+++ src/digisos/styles/digisos.less
@@ -2,7 +2,6 @@
     App spesic styling
 */
 
-@import "../informasjon/personopplysninger.less";
 @import "../skjema/ekstrainformasjon/vedlegg/slettIkon.less";
 @import "../skjema/ekstrainformasjon/vedlegg/vedleggsliste.less";
 @import "../skjema/familie/barn.less";
